requirements.txt
Flask
psycopg2-binary
python-docx
pymupdf
pytesseract
pillow